Update windows auto build script.ps1

+++ windows auto build script.ps1
@@ -1,7 +1,9 @@
 cd $HOME\Desktop
-if (Test-Path build_temp) { rm -Recurse -Force build_temp }
+if (Test-Path build_temp) { rm -Recurse -Force build_temp 
+if (Test-Path build_temp) { rm -rf build_temp }}
+
 mkdir build_temp; cd build_temp
-git clone https://github.com/rsa17826/textreplace-vscode-extension .
+git clone https://github.com/rsa17826/fixed-line-actions-vscode-extension.git .
 
 $nodeDir = "$HOME\Desktop\node-v22.0.0-win-x64"
 if (-not (Test-Path $nodeDir)) {
@@ -18,4 +20,5 @@ echo y|npx @vscode/vsce package --allow-missing-repository --no-git-tag-version
 
 mv *.vsix ..
 cd ..
-rm -Recurse -Force build_temp
+if (Test-Path build_temp) { rm -Recurse -Force build_temp 
+if (Test-Path build_temp) { rm -rf build_temp }}
